backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
python-multipart
pandas
scikit-learn
kaggle
